+++ src/main/kotlin/omok/model/stone/GoStoneColor.kt
@@ -1,4 +1,4 @@
-package omok.model
+package omok.model.stone
 
 enum class GoStoneColor {
     BLACK, WHITE;
