+++ src/test/kotlin/omok/model/CoordinateTest.kt
@@ -1,5 +1,6 @@
 package omok.model
 
+import omok.model.stone.Coordinate
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatIllegalArgumentException
 import org.junit.jupiter.api.Test
