+++ src/test/kotlin/omok/model/judge/OmokRuleTest.kt
@@ -1,11 +1,11 @@
 package omok.model.judge
 
 import omok.model.Board
-import omok.model.Coordinate
-import omok.model.GoStone
-import omok.model.GoStoneColor
 import omok.model.state.Stay
 import omok.model.state.Win
+import omok.model.stone.Coordinate
+import omok.model.stone.GoStone
+import omok.model.stone.GoStoneColor
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
